Reload when navigating between diff pages
By default github uses pushState w/ ajax which makes the content script
not run when changing pages. Also, using back/forward buttons is buggy
because we are shown the modified diff view (eg. side by side) but we
lose parsed diff state. Thus, in both of these cases, we force reload
the page. Need to eventually find a better way of handling this that's
more user-friendly